break out RequestTest and ResponseTest
